+++ CMakeLists.txt
@@ -60,7 +60,7 @@ add_subdirectory(libinterpreter)
 add_subdirectory(libevm)
 add_subdirectory(libp2p)
 add_subdirectory(libmptstate)
-add_subdirectory(libexecutivecontext)
+add_subdirectory(libexecutive)
 add_subdirectory(libblockverifier)
 add_subdirectory(libstorage)
 add_subdirectory(libstoragestate)
