+++ libblockverifier/BlockVerifierInterface.h
@@ -29,7 +29,7 @@
 #include <libethcore/Transaction.h>
 #include <libethcore/TransactionReceipt.h>
 #include <libevm/ExtVMFace.h>
-#include <libexecutivecontext/ExecutionResult.h>
+#include <libexecutive/ExecutionResult.h>
 #include <libmptstate/State.h>
 #include <memory>
 namespace dev
@@ -49,7 +49,8 @@ public:
     virtual ~BlockVerifierInterface(){};
 
     virtual ExecutiveContext::Ptr executeBlock(dev::eth::Block& block) = 0;
-    virtual std::pair<dev::eth::ExecutionResult, dev::eth::TransactionReceipt> executeTransaction(
+    virtual std::pair<dev::executive::ExecutionResult, dev::eth::TransactionReceipt>
+    executeTransaction(
         const dev::eth::BlockHeader& blockHeader, dev::eth::Transaction const& _t) = 0;
 };
