+++ libblockverifier/ExecutiveContext.cpp
@@ -23,9 +23,10 @@
 #include "ExecutiveContext.h"
 #include <libdevcore/easylog.h>
 #include <libethcore/Exceptions.h>
-#include <libexecutivecontext/ExecutionResult.h>
+#include <libexecutive/ExecutionResult.h>
 #include <libstorage/TableFactoryPrecompiled.h>
 
+using namespace dev::executive;
 using namespace dev::eth;
 using namespace dev::blockverifier;
 using namespace dev;
@@ -106,11 +107,11 @@ std::shared_ptr<storage::Table> ExecutiveContext::getTable(const Address& addres
     return tableFactoryPrecompiled->getmemoryTableFactory()->openTable(tableName);
 }
 
-std::shared_ptr<dev::eth::StateFace> ExecutiveContext::getState()
+std::shared_ptr<dev::executive::StateFace> ExecutiveContext::getState()
 {
     return m_stateFace;
 }
-void ExecutiveContext::setState(std::shared_ptr<dev::eth::StateFace> state)
+void ExecutiveContext::setState(std::shared_ptr<dev::executive::StateFace> state)
 {
     m_stateFace = state;
 }
