+++ libinitializer/Fake.h
@@ -149,10 +149,10 @@ public:
         usleep(1000 * (block.getTransactionSize()));
         return m_executiveContext;
     };
-    virtual std::pair<dev::eth::ExecutionResult, dev::eth::TransactionReceipt> executeTransaction(
-        const dev::eth::BlockHeader& blockHeader, dev::eth::Transaction const& _t)
+    virtual std::pair<dev::executive::ExecutionResult, dev::eth::TransactionReceipt>
+    executeTransaction(const dev::eth::BlockHeader& blockHeader, dev::eth::Transaction const& _t)
     {
-        dev::eth::ExecutionResult res;
+        dev::executive::ExecutionResult res;
         dev::eth::TransactionReceipt reciept;
         return std::make_pair(res, reciept);
     }
