+++ libledger/DBInitializer.cpp
@@ -29,6 +29,7 @@ using namespace dev::blockverifier;
 using namespace dev::db;
 using namespace dev::eth;
 using namespace dev::mptstate;
+using namespace dev::executive;
 
 namespace dev
 {
