+++ libledger/DBInitializer.h
@@ -25,7 +25,7 @@
 #include "LedgerParamInterface.h"
 #include <libblockverifier/ExecutiveContextFactory.h>
 #include <libdevcore/OverlayDB.h>
-#include <libexecutivecontext/StateFactoryInterface.h>
+#include <libexecutive/StateFactoryInterface.h>
 #include <libstorage/MemoryTableFactory.h>
 #include <libstorage/Storage.h>
 #include <memory>
@@ -93,7 +93,7 @@ private:
     std::shared_ptr<LedgerParamInterface> m_param;
     std::shared_ptr<dev::blockverifier::ExecutiveContextFactory> m_executiveContextFac;
     std::shared_ptr<dev::storage::MemoryTableFactory> m_memoryTableFac;
-    std::shared_ptr<dev::eth::StateFactoryInterface> m_stateFactory;
+    std::shared_ptr<dev::executive::StateFactoryInterface> m_stateFactory;
     dev::storage::Storage::Ptr m_storage = nullptr;
     dev::OverlayDB m_mptStateDB;
 };
