+++ libmptstate/MPTStateFactory.cpp
@@ -26,6 +26,7 @@
 using namespace dev;
 using namespace dev::mptstate;
 using namespace dev::eth;
+using namespace dev::executive;
 
 std::shared_ptr<StateFace> MPTStateFactory::getState()
 {
