+++ libmptstate/MPTStateFactory.h
@@ -26,7 +26,7 @@
 
 #include "MPTState.h"
 #include <libdevcore/OverlayDB.h>
-#include <libexecutivecontext/StateFactoryInterface.h>
+#include <libexecutive/StateFactoryInterface.h>
 #include <libstorage/Storage.h>
 #include <boost/filesystem.hpp>
 
@@ -34,7 +34,7 @@ namespace dev
 {
 namespace mptstate
 {
-class MPTStateFactory : public dev::eth::StateFactoryInterface
+class MPTStateFactory : public dev::executive::StateFactoryInterface
 {
 public:
     MPTStateFactory(u256 const& _accountStartNonce, boost::filesystem::path const& _basePath,
@@ -48,7 +48,7 @@ public:
         m_db = MPTState::openDB(m_basePath, m_genesisHash);
     };
     virtual ~MPTStateFactory(){};
-    virtual std::shared_ptr<dev::eth::StateFace> getState() override;
+    virtual std::shared_ptr<dev::executive::StateFace> getState() override;
 
 private:
     OverlayDB m_db;
