+++ test/unittests/libblockverifier/BlockVerifierTest.cpp
@@ -34,6 +34,7 @@ using namespace dev::eth;
 using namespace dev::storage;
 using namespace dev::blockverifier;
 using namespace dev::mptstate;
+using namespace dev::executive;
 
 namespace dev
 {
@@ -96,7 +97,7 @@ struct BlockVerifierFixture
     std::shared_ptr<FakeBlock> m_fakeBlock;
     dev::storage::LevelDBStorage::Ptr m_levelDBStorage;
     std::shared_ptr<leveldb::DB> m_db;
-    std::shared_ptr<dev::eth::StateFactoryInterface> m_stateFactory;
+    std::shared_ptr<dev::executive::StateFactoryInterface> m_stateFactory;
     std::unordered_map<Address, dev::eth::PrecompiledContract> m_precompiledContract;
 };
